+++ instrumentation/opentelemetry-instrumentation-botocore/src/opentelemetry/instrumentation/botocore/__init__.py
@@ -80,7 +80,7 @@ for example:
 
 import json
 import logging
-from typing import Collection
+from typing import Any, Collection, Dict, Optional, Tuple
 
 from botocore.client import BaseClient
 from botocore.endpoint import Endpoint
@@ -88,6 +88,9 @@ from botocore.exceptions import ClientError
 from wrapt import wrap_function_wrapper
 
 from opentelemetry import context as context_api
+from opentelemetry.instrumentation.botocore.extensions.types import (
+    _AwsSdkCallContext,
+)
 from opentelemetry.instrumentation.botocore.package import _instruments
 from opentelemetry.instrumentation.botocore.version import __version__
 from opentelemetry.instrumentation.instrumentor import BaseInstrumentor
@@ -97,7 +100,8 @@ from opentelemetry.instrumentation.utils import (
 )
 from opentelemetry.propagate import inject
 from opentelemetry.semconv.trace import SpanAttributes
-from opentelemetry.trace import SpanKind, get_tracer
+from opentelemetry.trace import get_tracer
+from opentelemetry.trace.span import Span
 
 logger = logging.getLogger(__name__)
 
@@ -157,12 +161,12 @@ class BotocoreInstrumentor(BaseInstrumentor):
         unwrap(Endpoint, "prepare_request")
 
     @staticmethod
-    def _is_lambda_invoke(service_name, operation_name, api_params):
+    def _is_lambda_invoke(call_context: _AwsSdkCallContext):
         return (
-            service_name == "lambda"
-            and operation_name == "Invoke"
-            and isinstance(api_params, dict)
-            and "Payload" in api_params
+            call_context.service == "lambda"
+            and call_context.operation == "Invoke"
+            and isinstance(call_context.params, dict)
+            and "Payload" in call_context.params
         )
 
     @staticmethod
@@ -182,97 +186,126 @@ class BotocoreInstrumentor(BaseInstrumentor):
         if context_api.get_value(_SUPPRESS_INSTRUMENTATION_KEY):
             return original_func(*args, **kwargs)
 
-        # pylint: disable=protected-access
-        service_name = instance._service_model.service_name
-        operation_name, api_params = args
+        call_context = _determine_call_context(instance, args)
+        if call_context is None:
+            return original_func(*args, **kwargs)
 
-        error = None
-        result = None
+        attributes = {
+            SpanAttributes.RPC_SYSTEM: "aws-api",
+            SpanAttributes.RPC_SERVICE: call_context.service_id,
+            SpanAttributes.RPC_METHOD: call_context.operation,
+            # TODO: update when semantic conventions exist
+            "aws.region": call_context.region,
+        }
 
         with self._tracer.start_as_current_span(
-            f"{service_name}", kind=SpanKind.CLIENT,
+            call_context.span_name,
+            kind=call_context.span_kind,
+            attributes=attributes,
         ) as span:
             # inject trace context into payload headers for lambda Invoke
-            if BotocoreInstrumentor._is_lambda_invoke(
-                service_name, operation_name, api_params
-            ):
-                BotocoreInstrumentor._patch_lambda_invoke(api_params)
+            if BotocoreInstrumentor._is_lambda_invoke(call_context):
+                BotocoreInstrumentor._patch_lambda_invoke(call_context.params)
 
-            self._set_api_call_attributes(
-                span, instance, service_name, operation_name, api_params
-            )
+            _set_api_call_attributes(span, call_context)
+            self._call_request_hook(span, call_context)
 
             token = context_api.attach(
                 context_api.set_value(_SUPPRESS_HTTP_INSTRUMENTATION_KEY, True)
             )
 
-            if callable(self.request_hook):
-                self.request_hook(
-                    span, service_name, operation_name, api_params
-                )
-
+            result = None
             try:
                 result = original_func(*args, **kwargs)
-            except ClientError as ex:
-                error = ex
+            except ClientError as error:
+                result = getattr(error, "response", None)
+                _apply_response_attributes(span, result)
+                raise
+            else:
+                _apply_response_attributes(span, result)
             finally:
                 context_api.detach(token)
 
-            if error:
-                result = error.response
-
-            if callable(self.response_hook):
-                self.response_hook(span, service_name, operation_name, result)
-
-            self._set_api_call_result_attributes(span, result)
-
-            if error:
-                raise error
+                self._call_response_hook(span, call_context, result)
 
             return result
 
-    @staticmethod
-    def _set_api_call_attributes(
-        span, instance, service_name, operation_name, api_params
+    def _call_request_hook(self, span: Span, call_context: _AwsSdkCallContext):
+        if not callable(self.request_hook):
+            return
+        self.request_hook(
+            span,
+            call_context.service,
+            call_context.operation,
+            call_context.params,
+        )
+
+    def _call_response_hook(
+        self, span: Span, call_context: _AwsSdkCallContext, result
     ):
-        if span.is_recording():
-            span.set_attribute("aws.operation", operation_name)
-            span.set_attribute("aws.region", instance.meta.region_name)
-            span.set_attribute("aws.service", service_name)
-            if "QueueUrl" in api_params:
-                span.set_attribute("aws.queue_url", api_params["QueueUrl"])
-            if "TableName" in api_params:
-                span.set_attribute("aws.table_name", api_params["TableName"])
+        if not callable(self.response_hook):
+            return
+        self.response_hook(
+            span, call_context.service, call_context.operation, result
+        )
 
-    @staticmethod
-    def _set_api_call_result_attributes(span, result):
-        if span.is_recording():
-            if "ResponseMetadata" in result:
-                metadata = result["ResponseMetadata"]
-                req_id = None
-                if "RequestId" in metadata:
-                    req_id = metadata["RequestId"]
-                elif "HTTPHeaders" in metadata:
-                    headers = metadata["HTTPHeaders"]
-                    if "x-amzn-RequestId" in headers:
-                        req_id = headers["x-amzn-RequestId"]
-                    elif "x-amz-request-id" in headers:
-                        req_id = headers["x-amz-request-id"]
-                    elif "x-amz-id-2" in headers:
-                        req_id = headers["x-amz-id-2"]
 
-                if req_id:
-                    span.set_attribute(
-                        "aws.request_id", req_id,
-                    )
+def _set_api_call_attributes(span, call_context: _AwsSdkCallContext):
+    if not span.is_recording():
+        return
 
-                if "RetryAttempts" in metadata:
-                    span.set_attribute(
-                        "retry_attempts", metadata["RetryAttempts"],
-                    )
+    if "QueueUrl" in call_context.params:
+        span.set_attribute("aws.queue_url", call_context.params["QueueUrl"])
+    if "TableName" in call_context.params:
+        span.set_attribute("aws.table_name", call_context.params["TableName"])
 
-                if "HTTPStatusCode" in metadata:
-                    span.set_attribute(
-                        SpanAttributes.HTTP_STATUS_CODE,
-                        metadata["HTTPStatusCode"],
-                    )
+
+def _apply_response_attributes(span: Span, result):
+    if result is None or not span.is_recording():
+        return
+
+    metadata = result.get("ResponseMetadata")
+    if metadata is None:
+        return
+
+    request_id = metadata.get("RequestId")
+    if request_id is None:
+        headers = metadata.get("HTTPHeaders")
+        if headers is not None:
+            request_id = (
+                headers.get("x-amzn-RequestId")
+                or headers.get("x-amz-request-id")
+                or headers.get("x-amz-id-2")
+            )
+    if request_id:
+        # TODO: update when semantic conventions exist
+        span.set_attribute("aws.request_id", request_id)
+
+    retry_attempts = metadata.get("RetryAttempts")
+    if retry_attempts is not None:
+        # TODO: update when semantic conventinos exists
+        span.set_attribute("retry_attempts", retry_attempts)
+
+    status_code = metadata.get("HTTPStatusCode")
+    if status_code is not None:
+        span.set_attribute(SpanAttributes.HTTP_STATUS_CODE, status_code)
+
+
+def _determine_call_context(
+    client: BaseClient, args: Tuple[str, Dict[str, Any]]
+) -> Optional[_AwsSdkCallContext]:
+    try:
+        call_context = _AwsSdkCallContext(client, args)
+
+        logger.debug(
+            "AWS SDK invocation: %s %s",
+            call_context.service,
+            call_context.operation,
+        )
+
+        return call_context
+    except Exception as ex:  # pylint:disable=broad-except
+        # this shouldn't happen actually unless internals of botocore changed and
+        # extracting essential attributes ('service' and 'operation') failed.
+        logger.error("Error when initializing call context", exc_info=ex)
+        return None